Better module structuring in order to allow reusing parts in other contexts
